lessaon_1_2 week2.React DOM and Components
