+++ src/Data.jsx
@@ -1,0 +1,47 @@
+
+const lists = [
+    {
+      title: "Apply course - Ruby on Rails",
+      url: "https://codethedream.org/apply/",
+      due: "October 6, 2024",
+      id: 0,
+    },
+    {
+      title: "Work on assignment 1 - Project Setup and React Basics",
+      url: "https://classes.codethedream.org/course/react-v2.1/ibis?week=1&lesson=Learning+to+Learn",
+      due: "November 5, 2024",
+      id: 1,
+    },
+    {
+      title: "Work on assignment 2 - React DOM And Components",
+      url: "https://classes.codethedream.org/course/react-v2.1/ibis?week=2&lesson=React+DOM+and+Components",
+      due: "November 12, 2024",
+      id: 2,
+    },
+    {
+      title: "Work on assignment 3 - Props State And Handlers",
+      url: "https://classes.codethedream.org/course/react-v2.1/ibis?week=3&lesson=Props+State+and+Handlers",
+      due: "November 19, 2024",
+      id: 3,
+    },
+    {
+      title: "Work on assignment 4 - Lifting State And Props Handling",
+      url: "https://classes.codethedream.org/course/react-v2.1/ibis?week=4&lesson=Lifting+State+and+Props+Handling",
+      due: "November 26, 2024",
+      id: 4,
+    },
+    {
+      title: "Work on assignment 5 - Hooks And Fragments",
+      url: "https://classes.codethedream.org/course/react-v2.1/ibis?week=6&lesson=Hooks+and+Fragments",
+      due: "December 10, 2024",
+      id: 5,
+    },
+    {
+      title: "Work on assignment 6 - Reusable Components Imperative React",
+      url: "https://classes.codethedream.org/course/react-v2.1/ibis?week=7&lesson=Reusable+Components+Imperative+React",
+      due: "December 17, 2024",
+      id: 6,
+    }
+  ];
+
+  export default lists;
